README.md for my TBG
This is a git commit that i'm making to my README.md file. The premise of this readme is to showcase my game, what it is as well as how it runs. It is also used to demonstrate my knowledge on a readme file and the proper markdown of it.